Fix 2 compiler warnings.
This commit fixes the following compiler warnings:

vacuum_ao.c:738:1: warning: ‘init_vacrelstats’ was used with no prototype before its definition [-Wmissing-prototypes]
  738 | init_vacrelstats()
      | ^~~~~~~~~~~~~~~~
CPartPruneStepsBuilder.cpp: In member function ‘PartitionedRelPruneInfo* gpdxl::CPartPruneStepsBuilder::CreatePartPruneInfoForOneLevel(gpdxl::CDXLNode*)’:
CPartPruneStepsBuilder.cpp:83:29: warning: comparison of integer expressions of different signedness: ‘gpos::ULONG’ {aka ‘unsigned int’} and ‘int’ [-Wsign-compare]
   83 |         for (ULONG i = 0; i < pinfo->nparts; ++i)
      |                           ~~^~~~~~~~~~~~~~~

+++ src/backend/gpopt/translate/CPartPruneStepsBuilder.cpp
@@ -82,7 +82,7 @@ CPartPruneStepsBuilder::CreatePartPruneInfoForOneLevel(CDXLNode *filterNode)
 	// partitions that survived static partition pruning; iterate over this list
 	// to populate pinfo->subplan_map, pinfo->relid_map & pinfo->present_parts
 	ULONG part_ptr = 0;
-	for (int i = 0; i < pinfo->nparts; ++i)
+	for (ULONG i = 0; (int) i < pinfo->nparts; ++i)
 	{
 		pinfo->subpart_map[i] = -1;
 		if (part_ptr < m_part_indexes->Size() &&
